+++ src/main/java/org/ivan/artshow/module/orderitem/service/OrderitemService.java
@@ -1,5 +1,7 @@
 package org.ivan.artshow.module.orderitem.service;
 
+import org.ivan.artshow.common.core.resultcode.ResultCodes;
+import org.ivan.artshow.common.exception.BizException;
 import org.ivan.artshow.module.orderitem.pojo.Orderitem;
 import org.ivan.artshow.module.orderitem.pojo.dto.OrderitemDTO;
 import org.ivan.artshow.module.orderitem.repository.OrderitemRepository;
@@ -24,6 +26,9 @@ public class OrderitemService implements IOrderitemService {
 
     @Override
     public Orderitem addOrderItem(OrderitemDTO orderItem){
+        if (orderItem == null) {
+            throw new BizException(ResultCodes.NULLPOINT);
+        }
         Orderitem norderItem = new Orderitem();
         BeanUtils.copyProperties(orderItem,norderItem);
         return orderItemRepository.save(norderItem);
@@ -31,19 +36,31 @@ public class OrderitemService implements IOrderitemService {
 
     @Override
     public void deleteOrderItem(Integer OrderItemId){
+        if (OrderItemId == null) {
+            throw new BizException(ResultCodes.NULLPOINT);
+        }
         orderItemRepository.deleteById(OrderItemId);
     }
 
     @Override
     public Orderitem updateOrderItem(OrderitemDTO OrderItem){
+        if (OrderItem == null) {
+            throw new BizException(ResultCodes.NULLPOINT);
+        }
         Integer OrderItemId = OrderItem.getOrderItemId();
-        Orderitem norderItem = orderItemRepository.findById(OrderItemId).orElseThrow(()-> new RuntimeException("要更新的物品不存在，ID：" + OrderItemId));
+        if (OrderItemId == null) {
+            throw new BizException(ResultCodes.NULLPOINT);
+        }
+        Orderitem norderItem = orderItemRepository.findById(OrderItemId).orElseThrow(()-> new BizException(ResultCodes.NOTFOUND));
         BeanUtils.copyProperties(OrderItem,norderItem);
         return orderItemRepository.save(norderItem);
     }
     @Override
     public Orderitem queryOrderItem(Integer orderItemId) {
-        return orderItemRepository.findById(orderItemId).get();
+        if (orderItemId == null) {
+            throw new BizException(ResultCodes.NULLPOINT);
+        }
+        return orderItemRepository.findById(orderItemId).orElseThrow(() -> new BizException(ResultCodes.NOTFOUND));
     }
 
 }
